markw: QCamera2: Set auto antibanding mode using country ISO code
This hardcodes a list of countries using a 60 Hz power grid, and sets the
antibanding mode based on whether or not the country reported by the
framework is in that list. There are far more 50 Hz countries than 60 Hz
countries, so the default setting when the country isn't known is 50 Hz.
This is also the reason why only the 60 Hz countries are enumerated (since
there are fewer of them).

Change-Id: Id77e4ff2c1e227cdc43bf31021caa13a44031614
Signed-off-by: NRanjan-17 <nalinishranjan05@gmail.com>